requirements.txt
annotated-types==0.7.0
anyio==4.5.2
blinker==1.8.2
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.1.8
colorama==0.4.6
distro==1.9.0
exceptiongroup==1.3.0
flask==3.0.3
flask-mail==0.10.0
flask-caching==2.1.0
flask-wtf==1.2.1
flask-sqlalchemy==3.0.5
flask-migrate==4.0.5
pymysql==1.1.0
cryptography==41.0.7
greenlet==3.1.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
importlib-metadata==8.5.0
itsdangerous==2.2.0
jinja2==3.1.6
jiter==0.9.1
MarkupSafe==2.1.5
openai==1.90.0
pyaes==1.6.1
aiohttp==3.10.8
pydantic==2.10.6
pydantic-core==2.27.2
python-dotenv==1.0.1
requests==2.32.4
sniffio==1.3.1
SQLAlchemy==2.0.27
tqdm==4.67.1
typing-extensions==4.13.2
urllib3==2.2.3
werkzeug==3.0.6
gunicorn==21.2.0
zipp==3.20.2
beautifulsoup4==4.13.4
lxml==4.9.3
pandas==2.0.3
undetected-chromedriver==3.5.5
Pillow==10.4.0
pytesseract==0.3.10
bleach==6.1.0 